+++ internal/models/microservice_status.go
@@ -12,6 +12,9 @@ type MicroserviceStatus struct {
 	PodID          string            `json:"podId,omitempty" yaml:"podId,omitempty"`
 	Percentage     float32           `json:"percentage" yaml:"percentage"`
 	ErrorMessage   *string           `json:"errorMessage,omitempty" yaml:"errorMessage,omitempty"`
+	LastError      string            `json:"lastError,omitempty" yaml:"lastError,omitempty"`
+	LastErrorAt    int64             `json:"lastErrorAt,omitempty" yaml:"lastErrorAt,omitempty"`
+	RestartCount   int               `json:"restartCount,omitempty" yaml:"restartCount,omitempty"`
 	IPAddress      *string           `json:"ipAddress,omitempty" yaml:"ipAddress,omitempty"`
 	ExecSessionIDs []string          `json:"execSessionIds,omitempty" yaml:"execSessionIds,omitempty"`
 	HealthStatus   *string           `json:"healthStatus,omitempty" yaml:"healthStatus,omitempty"`